crm/requirements.txt
# Django
Django==2.0.7
djangorestframework==3.8.2

# Auth
djangorestframework-simplejwt==3.2.3

# Swagger
drf-yasg==1.10.0

# Fake data
factory-boy==2.11.1

# Image
Pillow==5.2.0

# Database
psycopg2-binary==2.7.5

# Codestyle
pycodestyle==2.4.0

# Time
python-dateutil==2.7.3
pytz==2018.5